refactor(homepage): give every section band its own module
`SectionRenderer.tsx` held the dispatch AND the legacy middle-slot markup, so
every per-section change had to land in one file. That is fine for one change
and a merge-conflict factory for several: the section work queued behind this
touches six different bands.

The phase-aware CTA row and the three in-renderer views move out VERBATIM —
`PhaseCtaRow`, `FeaturedSpeakersSection`, `OrganizersSection`,
`ProgramHighlightsSection` — and the renderer keeps exactly what is global to
the page: the `_type` dispatch, the cancelled/archived short-circuit, and the
unknown-`_type` skip-with-warn. No markup, no props and no conditions changed;
the modules are server-safe (no `'use client'`, no module-scope context) so the
render tree the page reaches is the one it reached before.

Pure refactors are only worth as much as their proof, so the DOM-equality
snapshots in `SectionRenderer.test.tsx` were generated from the PRE-extraction
renderer and committed alongside: 10 snapshots covering the default
composition, both legacy bands with house and configured copy, their empty
states, and all four `PhaseCtaRow` branches (CFP, tickets, info + sold-out,
programme pre- and post-event). The extraction commit does not touch them and
they still match. The existing leaf stubs now surface their props as data
attributes, so the snapshots also pin what each band hands its leaf.
